Remove code for pandas and awkward
- Remove a bunch of now-dead code
- Update tests
- Some complex files are simple - so rename them.
